+++ docs/changelog.md
@@ -1,5 +1,6 @@
 2026/05/07 Release 4.1.6
 
+- Re-add support for the JRE 17 in matchbox-engine (#510)
 - Add validation statistics feature (#462)
 - Upgrade Spring Boot from 3.5.12 to 3.5.14 to fix predictable temp directory vulnerability (CVE-2026-40973)
 - Upgrade Thymeleaf from 3.1.4.RELEASE to 3.1.5.RELEASE to fix improper recognition of unauthorized syntax patterns (CVE-2026-40478)